refactor(ale_claw): Tier 4 — extract canonical_types.py (leaf module)
Moves the 6 content-block TypedDicts + ContentBlock union + CanonicalMessage
to a dependency-free canonical_types.py. canonical.py re-exports all 8 so every
`from .canonical import CanonicalMessage / ToolResultBlock / ...` (widely used
across context, agent_loop, replay, tests) is unchanged. This is the foundation
for splitting the sanitize pipeline next — both adapters and sanitize can import
the shared types from a leaf, avoiding cycles. Trimmed canonical.py's now-unused
Union/NotRequired/TypedDict imports (kept Literal — still used by TranscriptPolicy
+ canonical_to_anthropic_messages). Pure move; re-exports are identical objects;
full suite green (928 passed, 3 skipped).
